# PERF: DataFrame.transpose for pyarrow-backed