- [x] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [x] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [x] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [x] Added an entry in the latest `doc/source/whatsnew/v2.1.0.rst` file if fixing a bug or adding a new feature.

Similar to #52836, but for pyarrow-backed frames. 

```
import pandas as pd
import pyarrow as pa
import numpy as np

values = np.random.randn(1000, 1000)
df = pd.DataFrame(values, dtype=pd.ArrowDtype(pa.float64()))

%timeit df.T

# 346 ms ± 67.5 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)    -> main
# 42.5 ms ± 2.12 ms per loop (mean ± std. dev. of 7 runs, 10 loops each) -> PR
```

Renamed the masked implementation from `homogenous` -> `homogeneous`. Both may be valid(?), but `homogeneous` is used elsewhere.